FRONT END/requirements.txt
absl-py==2.3.1
accelerate==0.28.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
asgiref==3.9.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==14.4.0
bcrypt==4.3.0
blinker==1.9.0
branca==0.8.1
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cloudpickle==3.1.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
colorlog==6.9.0
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
ConfigSpace==1.2.1
contourpy==1.3.2
crypto==1.4.1
cryptography==45.0.5
cycler==0.12.1
dask==2025.7.0
dask-expr==2.0.0
dask-jobqueue==0.9.0
datasets==4.0.0
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1754523434/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
dill==0.3.8
distributed==2025.7.0
Django==5.2.6
django-ckeditor==6.7.3
django-js-asset==3.1.2
dnspython==2.8.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1745502089858/work
filelock==3.18.0
filetype==1.2.0
filterpy==1.4.5
Flask==3.1.1
Flask-Mail==0.10.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
flatbuffers==25.2.10
folium==0.20.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
func_timeout==4.3.5
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
gTTS==2.5.4
h5py==3.14.0
huggingface-hub==0.34.4
idna==3.7
ifaddr==0.2.0
imageio==2.37.0
imbalanced-learn==0.13.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1748711254/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
keras==3.11.2
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
lightgbm==4.6.0
lime==0.2.0.1
llvmlite==0.44.0
locket==1.0.0
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mysql-connector-python==9.4.0
mysqlclient==2.2.7
Naked==0.1.32
namex==0.1.0
narwhals==2.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.4.2
nltk==3.9.1
norfair==2.3.0
numba==0.61.2
numpy==1.25.2
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.1
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1733271261340/work
partd==1.4.2
patsy==1.0.1
pi_heif==1.1.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.3.0
pillow-avif-plugin==1.5.2
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
polars==1.32.3
pooch==1.8.2
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1744724089886/work
propcache==0.3.2
protobuf==6.31.1
psutil @ file:///D:/bld/psutil_1740663160591/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybboxes==0.1.6
pycparser==2.23
pycryptodome==3.23.0
pydeck==0.9.1
pyee==13.0.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
pylibsrtp==0.12.0
pyOpenSSL==25.3.0
pyparsing==3.2.3
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.1.1
pytorch-tabnet==4.1.0
pytz==2025.2
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1752564081/work
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1754238073265/work
referencing==0.36.2
regex==2025.9.1
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
roboflow==1.2.6
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.2.2
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.1
shap==0.48.0
shapely==2.1.1
shellescape==3.8.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sklearn-compat==0.1.3
slicer==0.0.8
smmap==5.0.2
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
statsmodels==0.14.5
stopit==1.1.2
streamlit==1.49.1
streamlit-webrtc==0.63.4
sympy==1.14.0
tabnet==0.1.6
tblib==3.1.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
terminaltables==3.1.10
tf_keras==2.20.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado @ file:///D:/bld/tornado_1754732031185/work
TPOT==1.1.0
tqdm==4.67.1
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
transformers==4.56.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
ultralytics==8.3.190
ultralytics-thop==2.0.16
update-checker==0.18.0
urllib3==2.5.0
v==1
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1733231326287/work
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.3
WTForms==3.2.1
xgboost==3.0.4
xxhash==3.5.0
xyzservices==2025.4.0
yarl==1.20.1
zict==3.0.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
